Rename basic_rtl to primitive
