+++ examples/ex03_proc/ProcDpathRTL.py
@@ -10,7 +10,7 @@ Author : Shunning Jiang
 
 from pymtl3 import *
 from pymtl3.stdlib.mem import mk_mem_msg
-from pymtl3.stdlib.basic_rtl import Adder, Incrementer, Mux, RegEn, RegEnRst, RegisterFile
+from pymtl3.stdlib.primitive import Adder, Incrementer, Mux, RegEn, RegEnRst, RegisterFile
 
 from .MiscRTL import AluRTL, ImmGenRTL
 from .TinyRV0InstRTL import OPCODE, RD, RS1, RS2, SHAMT
