+++ pymtl3/stdlib/ifcs/test/xcel_ifcs_test.py
@@ -8,7 +8,7 @@ Author : Yanghui Ou
 """
 
 from pymtl3 import *
-from pymtl3.stdlib.basic_rtl import RegisterFile
+from pymtl3.stdlib.primitive import RegisterFile
 from pymtl3.stdlib.ifcs import XcelMsgType, mk_xcel_msg
 from pymtl3.stdlib.queues import NormalQueueRTL
 from pymtl3.stdlib.test_utils import run_sim
