+++ pymtl3/stdlib/primitive/arithmetics.py
