+++ pymtl3/stdlib/primitive/crossbars.py
