+++ pymtl3/stdlib/primitive/encoders.py
