+++ pymtl3/stdlib/queues/enrdy_queues.py
@@ -9,7 +9,7 @@ Date   : Mar 9, 2018
 """
 
 from pymtl3 import *
-from pymtl3.stdlib.basic_rtl import Mux, Reg, RegEn, RegRst
+from pymtl3.stdlib.primitive import Mux, Reg, RegEn, RegRst
 from pymtl3.stdlib.ifcs import RecvIfcRTL, SendIfcRTL
