+++ pymtl3/stdlib/queues/valrdy_queues.py
@@ -1,5 +1,5 @@
 from pymtl3 import *
-from pymtl3.stdlib.basic_rtl import Mux, Reg, RegEn, RegisterFile
+from pymtl3.stdlib.primitive import Mux, Reg, RegEn, RegisterFile
 from pymtl3.stdlib.ifcs import InValRdyIfc, OutValRdyIfc
